refactor: move push to base class
